==> 03_Categories.ipynb <==
# Restructuring Categories as Binary per zip code

import pandas as pd 

# Let's import `west_df.csv` to create binary columns of shop categories

df = pd.read_csv('./datasets/west_df.csv')
df.head()

df.shape

# We're only focusing on `categories` in this notebook!

cats = ['category1', 'category2', 'category3', 'category4', 'category5', 'zip_code']
cat_df = df[cats]
cat_df.head()

#this code groups by zip code and collects all the unique categories for each column
categories_df = cat_df.astype(str).groupby('zip_code').agg(lambda x: ' '.join(x.unique()))

categories_df.head()

categories_df.shape

#combining all category columns together in order to turn into lists 
categories_df['all'] = categories_df['category1'] + ' ' + categories_df['category2'] + ' ' + categories_df['category3'] + ' ' + categories_df['category4'] + ' ' + categories_df['category5'] 

categories_df.head()

categories_df['all'][0].split(' ')

import numpy as np

# Getting unique categories for each zip code

categories_df['list_cat'] = categories_df['all'].str.split(' ').apply(np.unique)

categories_df.head()

#sanity check
categories_df['list_cat'][50]

# Creating an object containing unique categories of ALL zipcodes

unique_cats = categories_df['list_cat'].apply(pd.Series).stack().unique()

len(unique_cats)

type(unique_cats)

#sanity check 
'mexican' in unique_cats

for word in categories_df['list_cat'][0]:
    print(word)

# This code is **VERY** important
- This code creates a dictionary and if one of the 469 categories is in the unique list of categories per zip code, we will add 1 to the dictionary. If not, we will add 0. 

cat_dictionary = {}
def categorize(cell):
    for category in unique_cats:
        #for cell in text:
            #for word in cell:
        if category in cell:
            cat_dictionary[category] = 1
        else: 
            cat_dictionary[category] = 0
    return cat_dictionary

#creating new column from array to list 
categories_df['cat_list'] = categories_df['list_cat'].tolist()

#sanity check
categories_df.head(1)

#sanity check
categories_df['cat_list'][5]

#### Reading in zip codes file to append to future category dataframe since we need to append to the dataframe with avg latitude, longitude, and count of businesses based on price and ratings. 

main_zips = pd.read_csv('./datasets/MAINZIPCODES.CSV')
main_zips.head()

main_zips.shape

categories_df.shape

main_zips_list = list(main_zips['zipcode'])

features = ['category1', 'category2', 'category3', 'category4', 'category5', 'all', 'list_cat','cat_list']
categories_df = categories_df[features]

#here is the dataframe with the appended zip code
categories_df.reset_index(level=0, inplace=True)
categories_df.head()

#sanity
categories_df.shape

#sanity
len(main_zips_list)

categories_df.dtypes

#changing zip code column to numeric
categories_df["zip_code"] = pd.to_numeric(categories_df["zip_code"])

categories_df.dtypes

df = categories_df.loc[categories_df['zip_code'].isin(main_zips_list)]

df.head()

#sanity
df.shape

df['cat_list'][157]

#sanity
type(categorize(df['cat_list'][20]))

#sanity
len(df)

df.reset_index(inplace = True)

#saving dataframe for security
df.to_csv('./datasets/cat_sanitycheck.csv')

df.head()

len(df)

### This is the code where the magic happens! Remember that `categorize` definition? We're going to apply it!

the_dict = {}
for i in range(0, len(df)):
    unique_or_not = dict(categorize(df['cat_list'][i]))
    the_dict[i] = unique_or_not

the_cat_df = pd.DataFrame.from_dict(the_dict)

the_cat_df[25].sum()
#at category 25, there are 65 zip codes that have that category

the_cat_df.head()

the_cat_df.shape

the_cat_df.head()
#notice how the rows are categories and columns are the zip codes

# Finally transposed!

df = the_cat_df.T

df.head()

#### Removing `?` column

df.drop(labels = '?', axis = 'columns', inplace = True)

df.head()

df.shape

#saving it for security
df.to_csv('./datasets/category_binary_THEONE.csv')

# Sanity Checks

binary = pd.read_csv('./datasets/category_binary_THEONE.csv')
binary.head()

binary.shape

### Let's read in another dataframe to append the zip code and make sure we have the right binaries for each category!

lists = pd.read_csv('./datasets/cat_sanitycheck.csv')
lists.head()

lists.shape

binary.drop(labels = ['Unnamed: 0'], axis = 'columns', inplace = True)
binary.head()

#sanity
binary.head()

#sanity
lists.head()

### Reading in another dataframe to append the binary columns to construct the main dataframe

df = pd.read_csv('./datasets/NewSummarized.csv')
df.head()

#sanity
df.shape

#sanity
binary.shape

#sanity
lists.shape

#sanity
MOTHER = pd.concat([df, lists, binary], axis=1) 

#sanity
MOTHER.head()

#sanity
MOTHER.shape

# the sanity to my sanity
# nothing should be null 
MOTHER.isnull().sum().sum()

# Called `MOTHEROFMOTHERS.csv`

MOTHER.to_csv('./datasets/MOTHEROFMOTHERS.csv')
