==> 04_CheckingMother.ipynb <==
# **THE BIGGEST SANITY CHECK OF ALL TIME**
- This notebook will tell us if we summarized our zip codes correctly

import pandas as pd 

#reading in MOTHEROFMOTHERS
df = pd.read_csv('./datasets/MOTHEROFMOTHERS.csv')
df.head()

df.shape

df.isnull().sum().sum()

list(df.columns)

### Sanity Check
- We're going to see index 22 (aka zip code) and see all the unique business categories the zip code offers
- At the same time, we're going to return all columns where the cell value equals 1 (both of these lists should match)- excluding counts of businesses

df['cat_list'][22]

#checking specific row where cells ==1 (whihc returns column names)
df.columns[(df == 1).iloc[22]]

# They match!
- Note: We matched multiple rows to make sure the data is consistent

df.head()
