==> 07_CompilingScrapedIncome.ipynb <==
# This notebook will combine the potential features with the median we scraped 

import pandas as pd

mother = pd.read_csv('./datasets/MOTHEROFMOTHERS.csv')
mother.head()

all_zips = list(mother['zip_code'])

len(all_zips)

### Reading all the target variables we scraped

zip1 = pd.read_csv('./datasets/target1.csv')
zip2 = pd.read_csv('./datasets/target2.csv')
zip3 = pd.read_csv('./datasets/target3.csv')
zip4 = pd.read_csv('./datasets/target4.csv')
zip5 = pd.read_csv('./datasets/target5.csv')
zip6 = pd.read_csv('./datasets/target6.csv')

#dropping all Unnamed columns
zip1.drop(labels = 'Unnamed: 0', axis = 'columns', inplace= True)
zip2.drop(labels = 'Unnamed: 0', axis = 'columns', inplace= True)
zip3.drop(labels = 'Unnamed: 0', axis = 'columns', inplace= True)
zip4.drop(labels = 'Unnamed: 0', axis = 'columns', inplace= True)
zip5.drop(labels = 'Unnamed: 0', axis = 'columns', inplace= True)
zip6.drop(labels = 'Unnamed: 0', axis = 'columns', inplace= True)

#sanity
zip6.head()

#combining all 6 target variable DFs and calling it zips
frames = [zip1, zip2, zip3, zip4, zip5, zip6]

zips = pd.concat(frames)

#sanity check
zips.head()

#sanity check
zips.dtypes

#sanity check
zips.isnull().sum()

#sanity check
zips.shape

#sanity check
zips['zip'].nunique()

### Organizing list of zip codes to:
- List of zip codes where target variable is 0 (called `zero_zips`)
- List of zip codes where target variable is not zero (called `good_zips`)
- List of zip codes where it is missing target variables completely (called `missing_zips`)

#all zip codes that are zero 
zero_zips = zips[zips['income'] ==0]

good_zips = zips[zips['income'] != 0] 

len(zero_zips)

zero_zips_only = list(zero_zips['zip'])

good_zips_only = list(good_zips['zip'])

len(good_zips_only)

missing_zips = list(set(all_zips) - set(zero_zips_only)- set(good_zips_only))

len(missing_zips)

#sanity check
2299+69+55

#list of zip codes where we are missing our target variable 
missing_zips

#list of zip codes where target variable is 0
zero_zips_only

#sanity
69+55

#sanity
124+2299

#where we have the target variable for these zip codes
good_zips.shape

#sanity
2423-2299

#sanity
good_zips.head()

#sanity
good_zips.tail()

good_zips.reset_index(inplace = True)

good_zips.head()

good_zips.shape

#saving zip code and target variable where we scraped the income 
good_zips.to_csv('./datasets/GOODZIPS2299.CSV')

#sanity and seeing the min and max of median income... income disparity is a huge problem in America
# but this is capitalism 
good_zips['income'].describe()

good_zips[good_zips['income'] == 250000]
#in the heart of seattle

good_zips[good_zips['income'] > 140000]
#expensive neighorhoods will be talked about when going over feature selection
