==> 08_CreatingGrandma.ipynb <==
# Appending all columns together
- After manually finding median income, there are some zip codes where median income is 0
- We will drop these zip codes since it will create a bias with our features

import pandas as pd

df = pd.read_csv('./datasets/MOTHEROFMOTHERS.csv')
df.head()

df.shape

### This is the csv where we manually collected median income

manual_zips = pd.read_csv('./datasets/erin_zips_excel.csv', header = None)
manual_zips.head()

manual_zips.shape

manual_zips.columns = ['zip_code', 'income']

manual_zips.head()

### Bringing in the good zip codes that had the median income from the first phase of scraping

scrape_zips = pd.read_csv('./datasets/GOODZIPS2299.CSV')
scrape_zips.head()

#removing unnecessary columns
scrape_zips.drop(labels = ['Unnamed: 0', 'index'], axis= 'columns', inplace= True)

scrape_zips.shape

scrape_zips.columns = ['zip_code', 'income']

scrape_zips.head()

manual_zips.head()

#sanity
manual_zips.dtypes

#sanity
scrape_zips.dtypes

#joining dataframes 
frames = [manual_zips, scrape_zips]
all_zips_income = pd.concat(frames)

all_zips_income.head()

all_zips_income.shape

all_zips_income.reset_index(inplace= True)

all_zips_income.shape

len(all_zips_income[all_zips_income['income'] ==0])

#there are 50 zip codes where median is 0
#includes where both median and averages are 0
#or median is 0 and average is not 0 

# Zip Codes with Target Variable as 0
- We're going to drop them 
- Use these parameters as our demo dataset to predict income

#demo_zips is a df where all zip codes have median income of 0
demo_zips = all_zips_income[all_zips_income['income'] ==0]


demo_zips.head()

demo_zips.shape

#saving the demo zip codes 
demo_zips.to_csv('./datasets/demozipslist.csv')

df.dtypes

all_zips_income.dtypes

#merging all df where target variable is valid 
main_df = pd.merge(df, all_zips_income, left_on='zip_code', right_on='zip_code', how='left').drop(['index_y', 'index_x'], axis=1)

#sanity
main_df.isnull().sum().sum()

#sanity
main_df.head()

# Final:
- Our full dataset has 2423 rows

main_df.shape

main_df.drop(labels = ['Unnamed: 0', 'Unnamed: 0.1', 'Unnamed: 0.1.1'], axis = 'columns', inplace = True)

main_df.head()

all_columns = list(main_df.columns)

#dropping unnecessary columns
main_df.drop(labels = ['Unnamed: 0.2', 'zip_code.1'], axis = 'columns', inplace = True)

all_columns

main_df.shape

#sanity
468+4+13+11

#sanity (total zip codes - $0 zip codes)
2423-50

main_df['income'].dtype

#creating df where we have a valid median income
df_2373 = main_df[main_df['income'] != 0]


#sanity
df_2373.shape

#just want to see the poorest zip code
df_2373[df_2373['income'] <3000]

# Checking if rows have correct categories 

# Sanity Check part II
- To make sure everything is correct

df_2373.reset_index(inplace = True)

df_2373.head()

#this is the unique categories at index (aka zip code) 666
df_2373['cat_list'][666]

#checking specific row where cells ==1 (whihc returns column names)
df_2373.columns[(df_2373 == 1).iloc[666]]

#it matches!

df_2373['count'][666]
#number of businesses is accurate 

df_2373.loc[666:666]
#income is accurate
#number of business is accurate 

df_2373.drop(labels = 'index', axis = 'columns', inplace= True)

df_2373.shape

# Since we have the `MOTHEROFMOTHER.csv`, we're calling this dataframe `Grandma.csv`

df_2373.to_csv('./datasets/Grandma.csv')

# Now it's modeling time!
