==> 11_TheCreationOfDemos.ipynb <==
# The following code will transform the demo data to our modelng data
- Demo data will have the EXACT 110 features
- In the same order in orer to be fed into the Pickle on Flask 

import pandas as pd

df = pd.read_csv('./datasets/MOTHEROFMOTHERS.csv')
df.head()

df.shape

demo_zips = pd.read_csv('./datasets/demozipslist.csv')
demo_zips.head()

demo_zips.shape

demo_list = list(demo_zips['zip_code'])

# demo_df = df[df['zip_code'].isin(demo_list)]

# Feature engineering for Demos

df['price_1_percentage'] = df['Price_1']/ df['count']
df['price_2_percentage'] = df['Price_2']/ df['count']
df['price_3_percentage'] = df['Price_3']/ df['count']
df['price_4_percentage'] = df['Price_4']/ df['count']

df['rate_1_percentage'] = df['Rating_1']/ df['count']
df['rate_1.5_percentage'] = df['Rating_1.5']/ df['count']
df['rate_2_percentage'] = df['Rating_2']/ df['count']

df['rate_2.5_percentage'] = df['Rating_2.5']/ df['count']
df['rate_3_percentage'] = df['Rating_3']/ df['count']
df['rate_3.5_percentage'] = df['Rating_3.5']/ df['count']

df['rate_4_percentage'] = df['Rating_4']/ df['count']
df['rate_4.5_percentage'] = df['Rating_4.5']/ df['count']
df['rate_5_percentage'] = df['Rating_5']/ df['count']

main_df = pd.get_dummies(df, columns = ['city'])

main_df.shape

main_df.head()

len(demo_list)

demo_df = main_df[main_df['zip_code'].isin(demo_list)]

demo_df.shape

main_list = ['zip_code',
 'acaibowls',
 'asianfusion',
 'bakeries',
 'breakfast_brunch',
 'bubbletea',
 'cafes',
 'coffee',
 'cupcakes',
 'desserts',
 'french',
 'gluten_free',
 'greek',
 'indpak',
 'italian',
 'japanese',
 'mediterranean',
 'mideastern',
 'newamerican',
 'persian',
 'pizza',
 'salad',
 'sandwiches',
 'seafood',
 'sushi',
 'thai',
 'turkish',
 'vegetarian',
 'vietnamese',
 'wine_bars',
 'city_Alamo',
 'city_Bellevue',
 'city_Beverly Hills',
 'city_Carlsbad',
 'city_Corona Del Mar',
 'city_Cupertino',
 'city_Danville',
 'city_Fremont',
 'city_Irvine',
 'city_Issaquah',
 'city_Kensington',
 'city_La Cañada Flintridge',
 'city_Ladera Ranch',
 'city_Los Altos',
 'city_Los Gatos',
 'city_Malibu',
 'city_Manhattan Beach',
 'city_Menlo Park',
 'city_Mercer Island',
 'city_Moraga',
 'city_Moss Beach',
 'city_Newport Beach',
 'city_Orinda',
 'city_Pacific Palisades',
 'city_Palo Alto',
 'city_Pleasanton',
 'city_Port Costa',
 'city_Portola Valley',
 'city_Redwood City',
 'city_Rolling Hills Estate',
 'city_Ross',
 'city_Sammamish',
 'city_San Francisco',
 'city_San Jose',
 'city_San Marino',
 'city_San Mateo',
 'city_San Ramon',
 'city_Santa Clara',
 'city_Saratoga',
 'city_Seattle',
 'city_Sierra City',
 'city_Sunnyvale',
 'city_Tiburon',
 'city_Trabuco Canyon',
 'city_Villa Park',
 'city_Woodinville',
 'city_Woodside',
 'city_Yorba Linda',
 'price_2_percentage',
 'price_3_percentage',
 'rate_4_percentage',
 'casinos',
 'galleries',
 'salvadoran',
 'soulfood',
 'city_Avery',
 'city_Bodfish',
 'city_Clearlake',
 'city_Desert Hot Springs',
 'city_Detroit',
 'city_El Monte',
 'city_Fresno',
 'city_Hemet',
 'city_Inglewood',
 'city_Los Angeles',
 'city_Marblemount',
 'city_Miramonte',
 'city_Oroville',
 'city_San Bernardino',
 'city_Sierraville',
 'city_Skamokawa',
 'city_Slab City',
 'city_Spokane',
 'city_Stockton',
 'city_Sumpter',
 'city_Thermal',
 'city_Vernon',
 'city_Wilderville',
 'price_1_percentage',
 'rate_2_percentage',
 'count']

demo = demo_df[main_list]
demo.head()

demo.shape

demo.head()

demo.reset_index(inplace= True, drop= True)

demo.head()

demo.shape

demo.to_csv('./datasets/MARIOdemoMODEL.csv')

demo_df.head()

demo_df.shape

features_for_graphs = ['zip_code', 'Price_1', 'Price_2', 'Price_3', 'Price_4', 'count']
demo_graphs = demo_df[features_for_graphs]
demo_graphs.head()

demo_graphs.reset_index(inplace=True, drop = True)

demo_graphs.head()

demo_graphs.shape

demo_graphs.to_csv('./datasets/MARIOdemoGRAPHS.csv')
